clean-chat-back/pyproject.toml
[tool.poetry]
name = "clean-chat-back"
version = "0.1.0"
description = ""
authors = ["Vitaliy Vorobyov <vitalik.w555@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.86.0"
uvicorn = "^0.19.0"
pydantic = "^1.10.2"
SQLAlchemy = "1.4.41"
bcrypt = "^4.0.1"
PyJWT = "^2.6.0"
python-decouple = "^3.6"
websockets = "^10.4"
psycopg-binary = "^3.1.4"
email-validator = "^1.3.0"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
databases = {extras = ["asyncpg"], version = "^0.6.2"}
celery = "^5.2.7"
detoxify = "^0.5.0"
redis = "^4.3.4"
flower = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
